+++ lib/an_lifecycle_cancellable.dart
@@ -8,7 +8,6 @@ export 'dialog/navigator_ext.dart';
 export 'key/key.dart';
 export 'lifecycle/lifecycle_async.dart';
 export 'lifecycle/lifecycle_context_effect.dart';
-export 'lifecycle/lifecycle_context_remember.dart';
 export 'lifecycle/lifecycle_controller.dart';
 export 'lifecycle/lifecycle_data.dart';
 export 'lifecycle/lifecycle_effect.dart';