+++ lib/remember.dart
@@ -1,0 +1,4 @@
+@Deprecated('use package:remember')
+library remember;
+
+export 'lifecycle/lifecycle_context_remember.dart';